Adding consolidated error message when multiple fields are not defined at a given level (#1089)
If multiple fields are not defined at a given level (for example if you move from full pipeline execution to only executing a subset) you can get a pretty awful set of error messages. This consolidates to into a single error message that communicates all of the fields that are not defined.