Rework the Run UI so that there is no fullscreen loading state (#1226)
This PR allows the PipelineRunContainer to display with `run=undefined` before the run has been loaded. This means that the UI shows immediately and there aren't any full-screen loading indicators when opening a run.

Videos below are intentionally running ~6x slower than normal. The right sidebar width changing is just when the video loops. Unfortunately we can't get rid of the initial flicker when the tab is opened and the page hasn't loaded yet, but this makes it a lot less 'jumpy'

Edit: I've also removed the "No logs to display" loading state that appears briefly in the `After` video. Ideally dagit will be reliably fast enough that no loading state should be necessary there and having it be blank while the data is fetched should be fine. (And then subsequently it'd show "No logs" if the result was `[]`)

Before:
![Before-660](https://user-images.githubusercontent.com/1037212/56007550-00cc2100-5c8e-11e9-93d5-35547f759827.gif)

After:
![After-650](https://user-images.githubusercontent.com/1037212/56007567-06296b80-5c8e-11e9-92cc-8a95c41dcad2.gif)