Add EMR support (#1283)
This changeset adds an EMR solid, which is designed for job-scoped clusters. With this solid, you can spin up an EMR cluster, run a sequence of steps on that cluster, and then terminate it upon completion.

Note that more work will be required here to support long-lived clusters or clusters that run several steps/jobs. To support those cases, we plan to model EMR as a resource, but first we need to flush out solid-scoped vs. pipeliine-scoped resource provisioning.