Consolidate two DauphinStepEvent creation paths
Summary:
This eliminates the code duplication between step event creation in
startPipelineExecution and executePlan. This ends up being a bit janky right
because in the executePlan codepath we end up constructing the logger messages in
addition to the dagster events, using the event_callback to collect them, and then
passing them through the generic event factory method, and then filtering out those
log messages. This is the "minimum viable" change to consolidate the codepaths and
I think this is worth another look in terms of what it says about our eventing
codepaths.

Test Plan: Unit tests

Reviewers: alangenfeld, natekupp

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D41