Allow multi-modal pipelines
Summary:
This the first step to opting into multi-modal pipelines

This most precarious part of this is the refactoring of the guts
of the resource creation pipeline. It's a whole party. The core
of it is the ResourceCreationAdapter class that is special built
for this transition period. It handles the three different places
where resources can be created

Test Plan: Unit tests

Reviewers: max, alangenfeld

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D62