Reenable airflow tests
Summary:
This avoids a second call to cli.logs on the hunch that what we're seeing is some version of
the elusive, long-running, "Docker hangs when you write too much to STDOUT all at once" bug. We
also upgrade the buildkite docker plugin to try to eliminate any issues we might be facing based
on old Docker versions.

Test Plan: Run the unit tests a few times and don't observe stochastic failures

Reviewers: alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D72