Add mode to cli
Summary:
This adds mode to the CLI and basic tests for it

I'm also using this opportunity to start to consolidate
our test pipelines into a single top-level repo so that they can
be reused in our dagster tests and have a single place where
we can view them in dagit. Things got a bit out of control. We'll
also have collections of environments. This should also serve as a
form of useful documentation

Test Plan: Unit tests

Reviewers: alangenfeld, max

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D76