Fix crash in dagit when steps are skipped
Summary:
Dagit was crashing anytime that it encountered a skip event. It appears to happen because immer's
produce doesn't play totally nicely with typescript correctness. In this case "skipped" needs to be
treated like "start" and create a metadata structure from scratch. Immer allowed an IRunMetaDict struct
to be created without a displayEvents property, and that caused a crash in ExecutionPlanBox

Test Plan: Run pipeline with failure and skip event

Reviewers: max, alangenfeld, bengotow

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D119