(airline-demo-refactor-5) Add KeyedFileStore and mirror_keyed_file_from_s3
Summary:
This adds what I hope will be a broadly used set of
abstractions for handling files.

- A KeyedFileStore: This makes it straightforward to just tuck away
files somewhere during computation. So in local dev mode it is
a key-value store of files as in a flat directory. But when going
in aws it would be a key-value store of files in an s3 bucket.
- mirror_keyed_file_from_s3: This is a generic solid that mirrors
a file from an s3 bucket into your keyed store. In the file system case
this would download the file to the local filesystem. In the s3 case
it would download the file from the source s3 bucket to the local node,
and then upload it to target. See the description for more details.

The next step will be to write an unzip_solid that consumes a file
handle in the airline demo pipeline. In the local case this will
copy the input file into an emphemeral file, unzip it, and then pass
on a file handle (still TBD). In the s3 case this would download,
unzip the file locally, and then re-upload that file into the s3 keyed
store.

Test Plan: Local unit tests

Reviewers: max, natekupp

Reviewed By: natekupp

Differential Revision: https://dagster.phacility.com/D313