(storage-plugin-system-1) Fully Pluggable System Storage
Summary:
This diff seeks to replace our RunStorageMode enum with
hardcoded options for system storage (current in_memory, filesystem, and
s3) with a pluggable system. This cuts the dynamic, unexpressed
dependency between dagster and dagster_aws. This also allows for fully
pluggable storage so that we can easily add storage for gcp, azure, or
other customized

It accomplishes this by adding a new concept of a
SystemStorageDefinition. Mode definitions take a list of these
to indicate what are the storate systems are supported for that mode.

The SystemStorageDefinition defines a config section and a function
that returns a SystemStoragePluginData which currently contains
an instance of RunStorage and an IntermediatesManager. This will soon
expand to include a FileManager (or something akin to that).

This also allows one to define mixed storage definitions. E.g. the
intermediates manager could point to s3, and the runs storage could
point to dynamo or something. This is really flexible.

Questions to ask --

1) What is the relationship between these and resources? Are they just a special type of resource? Should you allow to depend on resources?  (e.g share the same s3 session object between user land and the framework)
2) Like resources, should this by a dictionary so you could, for example, instantiate multiple instances of a single system storage definition? I believe the answer is no but I wanted to surface this.
3) Is ModeDefinition the right place for this? Other possibility were at the pipeline level or at the repository level. Mode is the most flexible, so I started there. I think mode definitions should be reused a lot across a set of pipelines or repo so I think this is fine.

Next steps --

1) Add a @system_storage decorator
2) Write a generic file manager so that userland code has access to a scratch place to pass files in between solids during a particular run
3) Restructure the type plugin system with this. Totally eliminating RunStorageMode is dependent on this.
4) Eliminate RunStorageMode. Replace all references to storage_mode to "storage_key" or similar to avoid ambiguity with top-level mode.
5) Determine/explore relationship with resources.
6) Write up documentation/guide for this
7) Write a gcp plugin

Test Plan: Buildkite

Reviewers: alangenfeld, max, natekupp

Reviewed By: max, natekupp

Differential Revision: https://dagster.phacility.com/D329