(storage-plugin-system-8) S3FileManager
Summary:
This adds an S3FileManager and unit tests for it. This is
definitely a bit duplicative with KeyedFileStore and S3Resource but I
want to keep all those implementations decoupled for now until we pin
down their APIs with battle testing. Once we have confidence in the APIs
then we can probably extract some common abstract to share code between
them.

Test Plan: buildkite

Reviewers: natekupp, alangenfeld, max

Reviewed By: alangenfeld, max

Differential Revision: https://dagster.phacility.com/D346