Initial improvements to the solid sidebar
Summary:
This diff updates the Dagit "solid sidebar":

- More clearly differentiate the current invocation and the solid definition by splitting the sidebar apart into two sections.

- Add an "Invocations" tab which shows you all the invocations in the current pipeline and allows you to click to jump to another instance.

- Make the collapsible headings "sticky" so you can opt to permanently hide Invocations, for example, or jump between invocations without the view resetting.

- Font sizes in the sidebar are a bit more consistent.

This diff also addresses the issues we had with handle IDs vs solid names. The URL bar is now solid names, not solid handles:
http://localhost:3000/airline_demo_ingest_pipeline/explore/process_q2_market_data/s3_to_df/download_from_s3_to_bytes

This fixes an issue that prevented clicking to navigate to another solid in the sidebar. In general in Dagit, the methods that allow you to focus / expand, etc. a particular solid all take either a solid `name` or a solid `path`, which is generated by splitting a handle.

{F19553}

Test Plan: Give it a spin!

Reviewers: alangenfeld, schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D404