Events API
Summary:
This is a pretty significant change to the events API, starting
with ExpectationResult and Materialization. Their APIs are now very
similar to eachother (identical except for an additional success bool
on ExpectationResult). The notable additional are of the "metadata
entries" of which currently their are four types: value, path, url, and
json. When the user emits one of these:

```
yield Materialization(
    label='Table: {table_name}'.format(table_name=table_name),
    description=(
        'Persisted table {table_name} in database configured in the db_info resource.'
    ).format(table_name=table_name),
    metadata_entries=[
        EventMetadataEntry.text(label='Host', text=context.resources.db_info.host),
        EventMetadataEntry.text(label='Db', text=context.resources.db_info.db_name),
    ],
)
```

It should be very clear what will be rendered in dagit. I'm imagining a
two-level display hierarhcy where each metadata entry corresponds to a
row.

This should be easy to extend in the future. Want to add a histogram
rendered? Add an EventMetdataEntry for a collection of numbers, add a
GraphQL type, and then add frontend support.

I did change the GraphQL schema, but I shoehorned the new data into the
existing client-side APIs (thank you Typescript!)

Looking for feedback as this is a substantial change.

Addresses https://github.com/dagster-io/dagster/issues/1408
Addresses https://github.com/dagster-io/dagster/issues/1438

Test Plan: Run things in dagit and look at them. Buildkite

Reviewers: max, natekupp, alangenfeld, bengotow

Reviewed By: natekupp, bengotow

Differential Revision: https://dagster.phacility.com/D413