Pull refactoring/cleanup off experimental EP branch
Summary:
Hey folks - I did some cleanup of the existing execution plan on the branch where I was working on the experimental Gaant view. Just pulling the bits that can be merged now so that the new structured log viewer can build on these changes and not cause merge conflicts later.

There are only two slight visual changes on this branch - I fixed the placement of the divider and added a bit of padding in one place in the execution plan where they seemed to be missing.

Test Plan: JS snapshot tests updated

Reviewers: alangenfeld, schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D512