Testing guide and changes to facilitate the guide
Summary:
To included in here is a testing guide, partially finished, but
I'd like to get a pass from folks. This also checks in a series of code
snippets that in the guide and puts them under test.

This also adds an execute_solid function (previously named execute_solid
is now execute_solid_in_pipeline). execute_solid gets passed a
ModeDefinition and dynamically constructs a pipeline, which I think
is going to be an increasingly common thing here. This gets us closer to
a world you think of solids executing within an emphemeral, dynamic
environment, rather than more static pipelines.

Test Plan: Buildkite

Reviewers: natekupp, alangenfeld, max

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D498