Fix grqphql playground asset paths
Summary:
Grqphql playground assets were being copied to /webapp/vendor instead of /webapp/build/vendor. This fixes the copy to the right path.

Also removes /webapp entirely as a build step in rebuild_dagit so that we shouldn't have this issue again

Test Plan: open localhost:3333/grqphql and it loads

Reviewers: #ft, max

Reviewed By: #ft, max

Differential Revision: https://dagster.phacility.com/D635