fix column resizing in run view
Summary:
The HTML5 attribute `draggable` on a styled component container was causing the `mouseup` event to get swallowed up.  Renaming the prop (used to key dynamic styles) caused the mouse events to fire as expected.

Refactored some of the document listeners so that they are tied to the component
lifecycle, instead of to UI events.

Issue: https://github.com/dagster-io/dagster/issues/1637

Test Plan: Resized columns in Firefox/Safari/Chrome.

Reviewers: #ft, bengotow, alangenfeld

Reviewed By: bengotow

Subscribers: alangenfeld, max

Differential Revision: https://dagster.phacility.com/D792