Fix state update error in ConfigEditorModePicker
Summary:
Previously, the `ConfigEditorModePicker` component was throwing the following errors, since `onModeChange` was being called in the `render` method, and `onModeChange` called `setState`.

{F36726}

This diff moves the `onModeChange` call to the `componentDidMount` and `componentDidUpdate` lifecycle methods to fix the error.

Test Plan: Loaded dagit, switched pipelines, error was not thrown and mode picker behaved as expected

Reviewers: #ft, bengotow, prha

Reviewed By: #ft, prha

Subscribers: prha

Differential Revision: https://dagster.phacility.com/D825