Break up AWS CLI and add tests
Summary:
This puts the dagster-aws CLI tool in the dagster-aws library; conceivably we may eventually have similar tools for GCP, Azure, etc. so it seems like this is a reasonable place, but open to other perspectives here!

Otherwise, main goal of this diff is breaking apart the monolithic script and adding tests for at least some of it.

Test Plan:
- unit
- manual tested `pip install -e ...; dagster-aws init; dagster-aws shell`

Reviewers: #ft, alangenfeld

Reviewed By: #ft, alangenfeld

Subscribers: alangenfeld

Differential Revision: https://dagster.phacility.com/D843