Fix errors for repositories without scheduler
Summary: If a repository didn't have a scheduler defined and the `SCHEDULER` feature flag was enabled, an error would be thrown because `scheduler_type` would be the abc `Scheduler`. This diff replaces the `get_scheduler_type` method on `RepositoryDefinition` with `build_scheduler`, which checks whether `scheduler_type` is defined and instantiates the `scheduler` if it is.

Test Plan: unit

Reviewers: #ft, natekupp

Reviewed By: #ft, natekupp

Differential Revision: https://dagster.phacility.com/D892