add hooks to route standard out/err from a compute step to a file
Summary:
Creates a contextmanager to wrap any compute function execution and
routes all streams to the filesystem if configured.  By default, we store all
stdout/stderr logs to a subdirectory within the $DAGSTER_HOME directory. All of
the existing behavior is maintained if $DAGSTER_HOME is not set.

Test Plan:
Adds a toy example that allows both the multiprocess and inprocess
execution engines to pipe stdout/stderr to disk.

Reviewers: #ft, natekupp

Reviewed By: #ft, natekupp

Subscribers: alangenfeld, natekupp

Differential Revision: https://dagster.phacility.com/D893